Removes the ability to trigger an onFlush callback.
Ultimately this feature made no sense the way it was implemented. It
would call a function onFlush (which the caller of push has no idea
when a flush will happen by design), but if no flush occured the
function would not be called.

If this ever gets re implemented it would save those callbacks and call
them when the pushed items get flushed.

Until a real use-case comes up we will leave it all theoretical.